+++ overlook_hotel/src/main/java/com/example/overlook_hotel/controller/api/UserController.java
@@ -1,6 +1,7 @@
 package com.example.overlook_hotel.controller.api;
-import com.example.overlook_hotel.model.User;
-import com.example.overlook_hotel.service.UserService;
+import com.example.overlook_hotel.model.entity.User;
+import com.example.overlook_hotel.service.hotel.UserService;
+
 import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.*;
 import java.util.List;
